Refactor: converter 수정
