+++ src/main/java/EatPic/spring/domain/card/controller/SearchController.java
@@ -1,6 +1,6 @@
 package EatPic.spring.domain.card.controller;
 
-import EatPic.spring.domain.card.dto.response.CardResponse;
+import EatPic.spring.domain.card.dto.response.SearchResponseDTO;
 import EatPic.spring.domain.card.repository.SearchRepository;
 import EatPic.spring.domain.card.service.SearchServiceImpl;
 import EatPic.spring.global.common.BaseResponse;
@@ -23,12 +23,12 @@ public class SearchController {
     @Operation(summary = "탐색 탭에서 모든 유저 리스트 조회",
             description = "기본 탐색 탭 조회 시 호출되는 api")
     @GetMapping("")
-    public BaseResponse<CardResponse.GetCardListResponseDto> getAllUsersInSearch(@RequestParam(value = "limit", required = false, defaultValue = "10") int limit,
-                                                                                   @RequestParam(value = "cursor", required = false) Long cursor
-                                                                                   // 유저 관련 처리는 추후에..
+    public BaseResponse<SearchResponseDTO.GetCardListResponseDto> getAllUsersInSearch(@RequestParam(value = "limit", required = false, defaultValue = "10") int limit,
+                                                                                      @RequestParam(value = "cursor", required = false) Long cursor
+                                                                                      // 유저 관련 처리는 추후에..
                                                     ) {
-        CardResponse.GetCardListResponseDto result = searchService.getAllCards(limit, cursor);
-        return BaseResponse.onSuccess(result);          // 리턴 부분 제대로 작동하는지 확인하기!
+        SearchResponseDTO.GetCardListResponseDto result = searchService.getAllCards(limit, cursor);
+        return BaseResponse.onSuccess(result);    // 리턴 부분 제대로 작동하는지 확인하기!
     }
 
 //    @Operation(summary = "검색 범위가 전체인 경우에서 계정 검색", description = "전체 - 계정 검색 api")
