+++ src/main/java/EatPic/spring/domain/card/service/SearchService.java
@@ -1,11 +1,7 @@
 package EatPic.spring.domain.card.service;
 
-import EatPic.spring.domain.card.dto.response.CardResponse;
-import EatPic.spring.domain.card.entity.Card;
-
-import java.util.List;
+import EatPic.spring.domain.card.dto.response.SearchResponseDTO;
 
 public abstract class SearchService {
-    // public abstract List<Card> getAllCards(int limit, Long cursor);
-    public abstract CardResponse.GetCardListResponseDto getAllCards(int limit, Long cursor);
+    public abstract SearchResponseDTO.GetCardListResponseDto getAllCards(int limit, Long cursor);
 }
